+++ temba/api/v1/tests.py
@@ -22,7 +22,7 @@ from temba.flows.models import FlowLabel, FlowRun, RuleSet, ActionSet, Flow
 from temba.locations.models import BoundaryAlias
 from temba.msgs.models import Msg
 from temba.orgs.models import Language
-from temba.tests import TembaTest, AnonymousOrg
+from temba.tests import TembaTest, AnonymousOrg, matchers
 from temba.utils.dates import datetime_to_json_date
 from temba.values.models import Value
 from temba.api.models import APIToken
@@ -464,7 +464,11 @@ class APITest(TembaTest):
         self.assertEqual(run.is_active, True)
         self.assertEqual(run.is_completed(), False)
         self.assertEqual(run.path, [
-            {'node_uuid': flow.entry_uuid, 'arrived_on': '2015-08-25T11:09:30.088000+00:00'}
+            {
+                'uuid': matchers.UUID4String(),
+                'node_uuid': flow.entry_uuid,
+                'arrived_on': '2015-08-25T11:09:30.088000+00:00'
+            }
         ])
 
         # check flow stats
@@ -633,7 +637,22 @@ class APITest(TembaTest):
         self.assertEqual(run.is_active, True)
         self.assertEqual(run.is_completed(), False)
         self.assertEqual(run.path, [
-            {'node_uuid': color_prompt.uuid, 'arrived_on': '2015-08-25T11:09:30.088000+00:00'}
+            {
+                'uuid': matchers.UUID4String(),
+                'node_uuid': color_prompt.uuid,
+                'arrived_on': '2015-08-25T11:09:30.088000+00:00'
+            }
+        ])
+        self.assertEqual(run.events, [
+            {
+                'type': 'msg_created',
+                'created_on': matchers.ISODate(),
+                'step_uuid': run.path[0]['uuid'],
+                'msg': {
+                    'uuid': matchers.UUID4String(),
+                    'text': 'What is your favorite color?'
+                }
+            }
         ])
 
         # outgoing message for reply
@@ -698,10 +717,29 @@ class APITest(TembaTest):
         self.assertEqual(run.is_active, False)
         self.assertEqual(run.is_completed(), True)
         self.assertEqual(run.path, [
-            {'node_uuid': color_prompt.uuid, 'arrived_on': '2015-08-25T11:09:30.088000+00:00', 'exit_uuid': color_prompt.exit_uuid},
-            {'node_uuid': color_ruleset.uuid, 'arrived_on': '2015-08-25T11:11:30.088000+00:00', 'exit_uuid': orange_rule.uuid},
-            {'node_uuid': color_reply.uuid, 'arrived_on': '2015-08-25T11:13:30.088000+00:00', 'exit_uuid': color_reply.exit_uuid},
-            {'node_uuid': new_node['uuid'], 'arrived_on': '2015-08-25T11:15:30.088000+00:00'}
+            {
+                'uuid': matchers.UUID4String(),
+                'node_uuid': color_prompt.uuid,
+                'arrived_on': '2015-08-25T11:09:30.088000+00:00',
+                'exit_uuid': color_prompt.exit_uuid
+            },
+            {
+                'uuid': matchers.UUID4String(),
+                'node_uuid': color_ruleset.uuid,
+                'arrived_on': '2015-08-25T11:11:30.088000+00:00',
+                'exit_uuid': orange_rule.uuid
+            },
+            {
+                'uuid': matchers.UUID4String(),
+                'node_uuid': color_reply.uuid,
+                'arrived_on': '2015-08-25T11:13:30.088000+00:00',
+                'exit_uuid': color_reply.exit_uuid
+            },
+            {
+                'uuid': matchers.UUID4String(),
+                'node_uuid': new_node['uuid'],
+                'arrived_on': '2015-08-25T11:15:30.088000+00:00'
+            }
         ])
 
         # joe should have an urn now