+++ temba/flows/tests.py
@@ -9,18 +9,17 @@ import pytz
 import re
 import six
 import time
-from uuid import uuid4
+
 from datetime import timedelta
 from decimal import Decimal
-
 from django.utils.encoding import force_text
 from mock import patch
 from openpyxl import load_workbook
-
 from django.conf import settings
 from django.core.urlresolvers import reverse
 from django.test.utils import override_settings
 from django.utils import timezone
+from uuid import uuid4
 
 from temba.airtime.models import AirtimeTransfer
 from temba.api.models import WebHookEvent, WebHookResult, Resthook
@@ -31,7 +30,7 @@ from temba.ussd.models import USSDSession
 from temba.locations.models import AdminBoundary, BoundaryAlias
 from temba.msgs.models import Broadcast, Label, Msg, INCOMING, PENDING, WIRED, OUTGOING, FAILED
 from temba.orgs.models import Language, get_current_export_version
-from temba.tests import TembaTest, MockResponse, FlowFileTest, also_in_flowserver, skip_if_no_flowserver, matchers
+from temba.tests import TembaTest, MockResponse, FlowFileTest, also_in_flowserver, skip_if_no_flowserver, matchers, MigrationTest
 from temba.triggers.models import Trigger
 from temba.utils.dates import datetime_to_str
 from temba.utils.goflow import FlowServerException, get_client, serialize_contact
@@ -443,8 +442,30 @@ class FlowTest(TembaTest):
 
         # check the path for contact 1
         self.assertEqual(contact1_run.path, [
-            {'node_uuid': str(color_prompt.uuid), 'arrived_on': matchers.ISODate(), 'exit_uuid': str(color_prompt.exit_uuid)},
-            {'node_uuid': str(color_ruleset.uuid), 'arrived_on': matchers.ISODate()}
+            {
+                'uuid': matchers.UUID4String(),
+                'node_uuid': str(color_prompt.uuid),
+                'arrived_on': matchers.ISODate(),
+                'exit_uuid': str(color_prompt.exit_uuid)
+            },
+            {
+                'uuid': matchers.UUID4String(),
+                'node_uuid': str(color_ruleset.uuid),
+                'arrived_on': matchers.ISODate()
+            }
+        ])
+        self.assertEqual(contact1_run.events, [
+            {
+                'type': 'msg_created',
+                'created_on': contact1_msg.created_on.isoformat(),
+                'step_uuid': contact1_run.path[0]['uuid'],
+                'msg': {
+                    'uuid': str(contact1_msg.uuid),
+                    'text': "What is your favorite color?",
+                    'urn': 'tel:+250788382382',
+                    'channel': {'uuid': str(self.channel.uuid), 'name': 'Test Channel'}
+                }
+            }
         ])
 
         # test our message context
@@ -516,9 +537,58 @@ class FlowTest(TembaTest):
         self.assertFalse(Flow.find_and_handle(extra)[0])
 
         self.assertEqual(contact1_run.path, [
-            {'node_uuid': str(color_prompt.uuid), 'arrived_on': matchers.ISODate(), 'exit_uuid': str(color_prompt.exit_uuid)},
-            {'node_uuid': str(color_ruleset.uuid), 'arrived_on': matchers.ISODate(), 'exit_uuid': str(orange_rule.uuid)},
-            {'node_uuid': str(color_reply.uuid), 'arrived_on': matchers.ISODate()}
+            {
+                'uuid': matchers.UUID4String(),
+                'node_uuid': str(color_prompt.uuid),
+                'arrived_on': matchers.ISODate(),
+                'exit_uuid': str(color_prompt.exit_uuid)
+            },
+            {
+                'uuid': matchers.UUID4String(),
+                'node_uuid': str(color_ruleset.uuid),
+                'arrived_on': matchers.ISODate(),
+                'exit_uuid': str(orange_rule.uuid)
+            },
+            {
+                'uuid': matchers.UUID4String(),
+                'node_uuid': str(color_reply.uuid),
+                'arrived_on': matchers.ISODate(),
+            }
+        ])
+        self.assertEqual(contact1_run.events, [
+            {
+                'type': 'msg_created',
+                'created_on': contact1_msg.created_on.isoformat(),
+                'step_uuid': contact1_run.path[0]['uuid'],
+                'msg': {
+                    'uuid': str(contact1_msg.uuid),
+                    'text': "What is your favorite color?",
+                    'urn': 'tel:+250788382382',
+                    'channel': {'uuid': str(self.channel.uuid), 'name': 'Test Channel'}
+                }
+            },
+            {
+                'type': 'msg_received',
+                'created_on': incoming.created_on.isoformat(),
+                'step_uuid': contact1_run.path[1]['uuid'],
+                'msg': {
+                    'uuid': str(incoming.uuid),
+                    'text': "orange",
+                    'urn': 'tel:+250788382382',
+                    'channel': {'uuid': str(self.channel.uuid), 'name': 'Test Channel'}
+                }
+            },
+            {
+                'type': 'msg_created',
+                'created_on': reply.created_on.isoformat(),
+                'step_uuid': contact1_run.path[2]['uuid'],
+                'msg': {
+                    'uuid': str(reply.uuid),
+                    'text': "I love orange too! You said: orange which is category: Orange You are: 0788 382 382 SMS: orange Flow: color: orange",
+                    'urn': 'tel:+250788382382',
+                    'channel': {'uuid': str(self.channel.uuid), 'name': 'Test Channel'}
+                }
+            }
         ])
 
         # we should also have a result for this RuleSet
@@ -2074,15 +2144,19 @@ class FlowTest(TembaTest):
         self.assertTrue(response.context['has_flows'])
         self.assertIn('flow_type', response.context['form'].fields)
 
-        # add call channel
-        twilio = Channel.create(self.org, self.user, None, 'T', "Twilio", "0785553434", role="C",
-                                secret="56789", gcm_id="456")
-
+        # our default brand has all choice types
         response = self.client.get(reverse('flows.flow_create'))
-        self.assertTrue(response.context['has_flows'])
-        self.assertIn('flow_type', response.context['form'].fields)  # shown because of call channel
+        choices = [(Flow.FLOW, 'Messaging'), (Flow.USSD, 'USSD Messaging'), (Flow.VOICE, 'Phone Call'),
+                   (Flow.SURVEY, 'Surveyor')]
+        self.assertEqual(choices, response.context['form'].fields['flow_type'].choices)
 
-        twilio.delete()
+        # now configure our deployment to ignore USSD
+        branding = copy.deepcopy(settings.BRANDING)
+        branding['rapidpro.io']['flow_types'] = [Flow.FLOW, Flow.VOICE, Flow.SURVEY]
+        choices = [(Flow.FLOW, 'Messaging'), (Flow.VOICE, 'Phone Call'), (Flow.SURVEY, 'Surveyor')]
+        with override_settings(BRANDING=branding):
+            response = self.client.get(reverse('flows.flow_create'))
+            self.assertEqual(choices, response.context['form'].fields['flow_type'].choices)
 
         # create a new regular flow
         response = self.client.post(reverse('flows.flow_create'), dict(name='Flow', flow_type='F'), follow=True)
@@ -4566,9 +4640,21 @@ class SimulationTest(FlowFileTest):
         self.admin.save()
 
         post_data = dict(has_refresh=True, version="1")
+
+        # simulate from two different users
+        self.login(self.editor)
+        response = self.client.post(simulate_url, json.dumps(post_data), content_type="application/json")
+
         self.login(self.admin)
         response = self.client.post(simulate_url, json.dumps(post_data), content_type="application/json")
         json_dict = response.json()
+
+        # check the activity, we should only see simulation for ourselves
+        simulation = response.json()
+        for count in six.itervalues(simulation['activity']):
+            self.assertEqual(1, count)
+        for count in six.itervalues(simulation['visited']):
+            self.assertEqual(1, count)
 
         self.assertEqual(len(json_dict.keys()), 6)
         self.assertEqual(len(json_dict['messages']), 2)
@@ -4577,6 +4663,11 @@ class SimulationTest(FlowFileTest):
 
         post_data['new_message'] = "3"
         post_data['has_refresh'] = False
+
+        # previous steps without exits shouldn't blow up simulation stats
+        run = FlowRun.objects.filter(contact=Contact.get_test_contact(self.admin)).first()
+        del run.path[0]['exit_uuid']
+        run.save(update_fields=('path',))
 
         response = self.client.post(simulate_url, json.dumps(post_data), content_type="application/json")
         self.assertEqual(200, response.status_code)
@@ -4678,8 +4769,9 @@ class FlowsTest(FlowFileTest):
     @also_in_flowserver
     def test_simple(self):
         favorites = self.get_flow('favorites')
-        action_set1 = favorites.action_sets.order_by('y').first()
-        rule_set1 = favorites.rule_sets.order_by('y').first()
+        action_set1, action_set3, action_set3 = favorites.action_sets.order_by('y')[:3]
+        rule_set1, rule_set2 = favorites.rule_sets.order_by('y')[:2]
+        red_rule = rule_set1.rules[0]
 
         run, = favorites.start([], [self.contact])
 
@@ -4721,6 +4813,67 @@ class FlowsTest(FlowFileTest):
                 'input': "I like red\nhttp://example.com/test.jpg"
             }
         })
+        self.assertEqual(run.path, [
+            {
+                'uuid': matchers.UUID4String(),
+                'node_uuid': str(action_set1.uuid),
+                'arrived_on': matchers.ISODate(),
+                'exit_uuid': str(action_set1.exit_uuid)
+            },
+            {
+                'uuid': matchers.UUID4String(),
+                'node_uuid': str(rule_set1.uuid),
+                'arrived_on': matchers.ISODate(),
+                'exit_uuid': str(red_rule['uuid'])
+            },
+            {
+                'uuid': matchers.UUID4String(),
+                'node_uuid': str(action_set3.uuid),
+                'arrived_on': matchers.ISODate(),
+                'exit_uuid': str(action_set3.exit_uuid)
+            },
+            {
+                'uuid': matchers.UUID4String(),
+                'node_uuid': str(rule_set2.uuid),
+                'arrived_on': matchers.ISODate()
+            }
+        ])
+        self.assertEqual(run.events, [
+            {
+                'type': 'msg_created',
+                'created_on': matchers.ISODate(),
+                'step_uuid': run.path[0]['uuid'],
+                'msg': {
+                    'uuid': str(msg1.uuid),
+                    'text': 'What is your favorite color?',
+                    'urn': 'tel:+12065552020',
+                    'channel': {'uuid': str(self.channel.uuid), 'name': 'Test Channel'}
+                }
+            },
+            {
+                'type': 'msg_received',
+                'created_on': matchers.ISODate(),
+                'step_uuid': run.path[1]['uuid'],
+                'msg': {
+                    'uuid': str(msg2.uuid),
+                    'text': 'I like red',
+                    'attachments': ['image/jpeg:http://example.com/test.jpg'],
+                    'urn': 'tel:+12065552020',
+                    'channel': {'uuid': str(self.channel.uuid), 'name': 'Test Channel'}
+                }
+            },
+            {
+                'type': 'msg_created',
+                'created_on': matchers.ISODate(),
+                'step_uuid': run.path[2]['uuid'],
+                'msg': {
+                    'uuid': matchers.UUID4String(),
+                    'text': 'Good choice, I like Red too! What is your favorite beer?',
+                    'urn': 'tel:+12065552020',
+                    'channel': {'uuid': str(self.channel.uuid), 'name': 'Test Channel'}
+                }
+            }
+        ])
 
         cat_counts = list(FlowCategoryCount.objects.order_by('id'))
         self.assertEqual(len(cat_counts), 1)
@@ -4810,11 +4963,11 @@ class FlowsTest(FlowFileTest):
             self.assertEqual(dict(uuid=six.text_type(run.uuid), created_on=run.created_on.isoformat()), payload['run'])
 
             # make sure things don't sneak into our path format unintentionally
-            # first item in the path should have node, arrived, and exit
-            self.assertEqual(3, len(payload['path'][0]))
+            # first item in the path should have uuid, node, arrived, and exit
+            self.assertEqual(set(payload['path'][0].keys()), {'uuid', 'node_uuid', 'arrived_on', 'exit_uuid'})
 
             # last item has the same, but no exit
-            self.assertEqual(2, len(payload['path'][-1]))
+            self.assertEqual(set(payload['path'][-1].keys()), {'uuid', 'node_uuid', 'arrived_on'})
 
             for key, value in six.iteritems(results):
                 result = payload['results'].get(key)
@@ -5671,7 +5824,7 @@ class FlowsTest(FlowFileTest):
         })
 
         # but hammer should have created some simulation activity
-        (active, visited) = flow.get_activity(simulation=True)
+        (active, visited) = flow.get_activity(hammer)
         self.assertEqual(active, {})
         self.assertEqual(visited, {
             '%s:%s' % (color_question.exit_uuid, color.uuid): 1,
@@ -8206,7 +8359,7 @@ class FlowBatchTest(FlowFileTest):
         stopped.stop(self.admin)
 
         # start our flow, this will take two batches
-        with QueryTracker(assert_query_count=308, stack_count=10, skip_unique_queries=True):
+        with QueryTracker(assert_query_count=298, stack_count=10, skip_unique_queries=True):
             flow.start([], contacts)
 
         # ensure 11 flow runs were created
@@ -8942,7 +9095,7 @@ class QueryTest(FlowFileTest):
 
         # mock our webhook call which will get triggered in the flow
         self.mockRequest('GET', '/ip_test', '{"ip":"192.168.1.1"}', content_type='application/json')
-        with QueryTracker(assert_query_count=140, stack_count=10, skip_unique_queries=True):
+        with QueryTracker(assert_query_count=138, stack_count=10, skip_unique_queries=True):
             flow.start([], [self.contact])
 
 
@@ -9283,3 +9436,93 @@ class AssetServerTest(TembaTest):
                 }
             ],
         })
+
+
+class BackfillRunEventsTest(MigrationTest):
+    migrate_from = '0149_update_path_trigger'
+    migrate_to = '0150_populate_run_events'
+    app = 'flows'
+
+    def setUpBeforeMigration(self, apps):
+        self.contact1 = self.create_contact("Joe", number='+250783835555')
+        self.contact2 = self.create_contact("Frank", number='+250783836666')
+        self.flow = self.get_flow('favorites')
+
+        contact1_run, contact2_run = self.flow.start([], [self.contact1, self.contact2])
+
+        def remove_step_uuid(path_step):
+            return {k: v for k, v in six.iteritems(path_step) if k != 'uuid'}
+
+        # make run #1 look like it was created before we started saving step uuids and msg events
+        contact1_run.path = [remove_step_uuid(s) for s in contact1_run.path]
+        contact1_run.events = []
+        contact1_run.save(update_fields=('path', 'events'))
+
+        squash_flowruncounts()
+        squash_flowpathcounts()
+
+    def test_events_created(self):
+        action_set1, action_set3, action_set3 = self.flow.action_sets.order_by('y')[:3]
+        rule_set1, rule_set2 = self.flow.rule_sets.order_by('y')[:2]
+
+        contact1_run, contact2_run = FlowRun.objects.order_by('contact__id')
+        contact1_msgs = Msg.objects.filter(contact=self.contact1).order_by('id')
+        contact2_msgs = Msg.objects.filter(contact=self.contact2).order_by('id')
+
+        self.assertEqual(contact1_run.path, [
+            {
+                'uuid': matchers.UUID4String(),
+                'node_uuid': str(action_set1.uuid),
+                'arrived_on': matchers.ISODate(),
+                'exit_uuid': str(action_set1.exit_uuid),
+            },
+            {
+                'uuid': matchers.UUID4String(),
+                'node_uuid': str(rule_set1.uuid),
+                'arrived_on': matchers.ISODate()
+            }
+        ])
+        self.assertEqual(contact1_run.events, [
+            {
+                'type': 'msg_created',
+                'created_on': contact1_msgs[0].created_on.isoformat(),
+                'msg': {
+                    'uuid': str(contact1_msgs[0].uuid),
+                    'channel': {'uuid': self.channel.uuid, 'name': 'Test Channel'},
+                    'text': 'What is your favorite color?',
+                    'urn': 'tel:+250783835555',
+                },
+                'step_uuid': contact1_run.path[0]['uuid']
+            }
+        ])
+
+        self.assertEqual(contact2_run.path, [
+            {
+                'uuid': matchers.UUID4String(),
+                'node_uuid': str(action_set1.uuid),
+                'arrived_on': matchers.ISODate(),
+                'exit_uuid': str(action_set1.exit_uuid),
+            },
+            {
+                'uuid': matchers.UUID4String(),
+                'node_uuid': str(rule_set1.uuid),
+                'arrived_on': matchers.ISODate()
+            }
+        ])
+        self.assertEqual(contact2_run.events, [
+            {
+                'type': 'msg_created',
+                'created_on': contact2_msgs[0].created_on.isoformat(),
+                'msg': {
+                    'uuid': str(contact2_msgs[0].uuid),
+                    'channel': {'uuid': self.channel.uuid, 'name': 'Test Channel'},
+                    'text': 'What is your favorite color?',
+                    'urn': 'tel:+250783836666',
+                },
+                'step_uuid': contact2_run.path[0]['uuid']
+            }
+        ])
+
+        # check that modifying paths didn't create extra counts
+        self.assertEqual(FlowPathCount.objects.count(), 1)
+        self.assertEqual(FlowNodeCount.objects.count(), 2)